# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: string;
  name: string;
  setName: string;
  number: string;
  imageUrl?: string | null;
}