# lib/api-zod/src/generated/types/detection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Detection {
  id: string;
  cardId?: string | null;
  cardName: string;
  setName: string;
  cardNumber: string;
  imageUrl?: string | null;
  detectedPrice?: number | null;
  rawPrice?: number | null;
  psa9Price?: number | null;
  psa10Price?: number | null;
  confidence: number;
  detectedAt: Date;
  source: string;
  duplicate: boolean;
}