# lib/api-zod/src/generated/types/identifyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IdentifyInput {
  cardName?: string | null;
  setName?: string | null;
  cardNumber?: string | null;
  imageBase64?: string | null;
}