# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  discordEnabled: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumConfidence: number;
  /**
     * @minimum 1
     * @maximum 60
     */
  scanInterval: number;
  /**
     * @minimum 1
     * @maximum 3600
     */
  duplicateCooldown: number;
  backendUrl: string;
}